As per SETUP.md
According to https://github.com/exercism/xpython/blob/master/SETUP.md, the solution should be in a file of the same name as the problem, this is not the case here.

There's an additional problem that this change does not solve: across exercism.io, exercises are named in kebab-case, whereas Python convention favours snake_case. This means the SETUP.md instructions are still not completely accurate. I'm not sure what's the best way to solve this.